Revert "Fix "Package is supported" label translation" (rgiconlegend-supported.patch)
If a user supplies his own text in the configuration, there is no hope
a translation would be found among the compiled ones, so that fix was
evil in general.

(But there has been a problem: since the packaged config file includes
a custom value for Synaptic::supported-text, without that fix, the
label would be shown untranslated in the UI; i.e., the custom text
would override any looked up compiled translations. However, that fix
would actually help--but only because the custom text in the config
file matched the default hardcoded label...)

I envision another better fix: just rely on the default value for most
cases by removing this string from the config. (See the next commit.)

This reverts:

    commit 34d0e974736a084bd1e3abd6bf2f56615ee2595b
    Author: Aleksei Nikiforov <darktemplar@altlinux.org>
    Date:   Thu Jun 13 15:52:49 2019 +0300

        Apply synaptic-0.58-rgiconlegend-supported.patch

initially introduced (as a patch file) in:

    commit e25bc10373311c5e55595cc6ca313c995cd4e564
    Author: Lenar Shakirov <snejok@altlinux.org>
    Date:   Thu Oct 6 23:30:03 2011 +0400

        Fix "Package is supported" label translation

Fixes: e25bc10 Fix "Package is supported" label translation
Fixes: 34d0e97 Apply synaptic-0.58-rgiconlegend-supported.patch