gtkpkglist.cc: fix some uses of c_str() value after destruction (ALT#40010)
We rewrite a piece of code in a simpler way (from a certain point of
view): it was not really needed to save the pointers first to an extra
local variable before use; instead, now we immediately pass the
pointer to the function that uses it.

Thereby we avoid the error of using the c_str() value after its owner
(a temporary object) was destroyed: the lifespan of a temporary is
until the full expression is evaluated; and that lifespan is enough
here since g_value_set_string()'s argument is a "caller-owned string
to be duplicated" (per the gobject docs).

(In general, there is no way to safeguard the code from errors with
unowned dangling C pointers by means of the type system or the like
during compilation: we don't know what lifespan of the pointer is
assumed by a function that takes it as an argument; and, unlike in the
case of g_value_set_string(), a function might assign the pointer to a
long-term variable instead of duplicating its content...)

Remarks:

Note that checking the c_str() value for NULL in one of the cases was
useless, since c_str() must return a pointer to a null-terminated
sequence of chars; thus we rewrite it in a simpler way without
temporary variable to hold the pointer.

Also checking pkg for NULL was bogus, because it's already done in the
beginning of the function body.

Reported-by: Alexander Makeenkov <amakeenk@basealt.ru>