adapt to the new ALT's libapt's (internal) API for custom callbacks
The code that needed to be adapted is actually a hack, because it uses
an internal API from libapt, which is declared "protected" and
actually shouldn't be used anywhere except for the library's
implementations of the abstract pkgPackageManager interafce. The
public method DoInstall() is intended for public use instead. But
synaptic needs to re-implement DoInstall() semantics just the same way
as in libapt, but with a fork in the middle...

Of course, although we might sometimes keep compatibility methods in apt
when we change the API in ALT, in this case, clearly, we shouldn't clutter
apt source code with a compat method to serve a hack in another project:
the compat method would have no justification from the point of view
of the logic behind apt's API.