adapt the dialog to show changelogs for ALT (they are in the index) & enable it
Apart from a new feature for users, this will also allow us to get rid of
pkgAcqFileSane class used to download the changelogs in Debian and
to recompile synaptic with any pkgAcquire::Item API in ALT's libapt.

Find it in the top menu: "Package" -> "Download Changelog". (TODO/patches
are welcome: Place it along the package properties like dependencies.)

TODO: I don't quite like the current use of static buffers in
functions like RPackage::maintainer(), which is the prototype for my new
RPackage::changelog(). But I changed it to returning a copy of std::string.

TODO: adapt and enable the display of some other info like vendor and
the list of the files of a package etc.